SignLink.API/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
numpy==1.25.0
pillow==10.4.0
h5py==3.8.0
opencv-python==4.10.0.84
mediapipe==0.10.5
protobuf==3.20.3
inference-sdk